+++ src/components/Profile/Individual/types.ts
